fix: 修复CI mypy/format错误，新增7个测试文件类型注解
## 修复内容
- st_cache.py: ttl参数类型 int→float（支持50ms等小数TTL测试）
- test_st_cache.py: clear()和__doc__添加type: ignore注释
- test_globals.py: 单例属性添加type: ignore注释
- test_experiment_registry.py: get()返回None时添加is not None检查
- test_unified_decision.py: prob_win/prob_loss为None时添加类型守卫
- test_eval_benchmark.py: get_best_alpha返回None时添加检查

## 影响
- 新增测试文件mypy错误从22降至0
- 整体mypy错误从174降至约165（遗留问题未动）

+++ trade_krono_cli/utils/st_cache.py
@@ -18,12 +18,12 @@ from typing import Any, TypeVar
 F = TypeVar("F", bound=Callable[..., Any])
 
 
-def cached(ttl: int = 1800) -> Callable[[F], F]:
+def cached(ttl: float = 1800) -> Callable[[F], F]:
     """装饰器：为函数结果添加 TTL 缓存。
 
     Parameters
     ----------
-    ttl : int
+    ttl : float
         缓存存活时间（秒），默认 1800（30 分钟）
 
     """
